==> src/01_preparation/1_read_webpages.R <==
setwd("Z:\\Respondi\\RESPONDI_w3\\")  

library(readxl)
Book1 <- read_excel("./splitfiles/Book1.xlsx")
Book2 <- read_excel("./splitfiles/Book2_2.xlsx")
Book3 <- read_excel("./splitfiles/Book3.xlsx")
all_data <- rbind(Book1 , Book2)
rm(Book1)
all_data <- rbind(all_data , Book3)
rm(Book2)
rm(Book3)

Book4 <- read_excel("./splitfiles/Book4_2.xlsx")
Book5 <- read_excel("./splitfiles/Book5_2.xlsx")
Book6 <- read_excel("./splitfiles/Book6.xlsx")
all_data <- rbind(all_data , Book4)
rm(Book4)
all_data <- rbind(all_data , Book5)
rm(Book5)
all_data <- rbind(all_data , Book6)
rm(Book6)

Book7 <- read_excel("./splitfiles/Book7_2.xlsx")
Book8 <- read_excel("./splitfiles/Book8_2.xlsx")
Book9 <- read_excel("./splitfiles/Book9_2.xlsx")
all_data <- rbind(all_data , Book7)
rm(Book7)
all_data <- rbind(all_data , Book8)
rm(Book8)
all_data <- rbind(all_data , Book9)
rm(Book9)

Book10 <- read_excel("./splitfiles/Book10_2.xlsx")
Book11 <- read_excel("./splitfiles/Book11_2.xlsx")
Book12 <- read_excel("./splitfiles/Book12_2.xlsx")
all_data <- rbind(all_data , Book10)
rm(Book10)
all_data <- rbind(all_data , Book11)
rm(Book11)
all_data <- rbind(all_data , Book12)
rm(Book12)

Book13 <- read_excel("./splitfiles/Book13_2.xlsx")
Book14 <- read_excel("./splitfiles/Book14_2.xlsx")
Book15 <- read_excel("./splitfiles/Book15_2.xlsx")
all_data <- rbind(all_data , Book13)
rm(Book13)
all_data <- rbind(all_data , Book14)
rm(Book14)
all_data <- rbind(all_data , Book15)
rm(Book15)

Book16 <- read_excel("./splitfiles/Book16_2.xlsx")
Book17 <- read_excel("./splitfiles/Book17_2.xlsx")
Book18 <- read_excel("./splitfiles/Book18_2.xlsx")
all_data <- rbind(all_data , Book16)
rm(Book16)
all_data <- rbind(all_data , Book17)
rm(Book17)
all_data <- rbind(all_data , Book18)
rm(Book18)

Book19 <- read_excel("./splitfiles/Book19_2.xlsx")
Book20 <- read_excel("./splitfiles/Book20_2.xlsx")
Book21 <- read_excel("./splitfiles/Book21_2.xlsx")
all_data <- rbind(all_data , Book19)
rm(Book19)
all_data <- rbind(all_data , Book20)
rm(Book20)
all_data <- rbind(all_data , Book21)
rm(Book21)

Book22 <- read_excel("./splitfiles/Book22_2.xlsx")
Book23 <- read_excel("./splitfiles/Book23_2.xlsx")
Book24 <- read_excel("./splitfiles/Book24_2.xlsx")
Book25 <- read_excel("./splitfiles/Book25_2.xlsx")
all_data <- rbind(all_data , Book22)
rm(Book22)
all_data <- rbind(all_data , Book23)
rm(Book23)
all_data <- rbind(all_data , Book24)
rm(Book24)
all_data <- rbind(all_data , Book25)
rm(Book25)

saveRDS(all_data, file = "./original daten/web_pageviews.rds")
